Serialize XMSS leaf index as four bytes
Internally XMSS uses a 64 Bit type for the leaf index.
This patch removes the four leading zero bytes from the
XMSS leaf index and serializes it as a four byte value
as described in RFC 8391.

Test cases are adjusted accordingly.

The 64 Bit type is kept internally which potentially
allows for code reuse when implementing XMSS^MT on top
of the current XMSS code.

+++ src/lib/pubkey/xmss/xmss_privatekey.h
@@ -206,7 +206,7 @@ class BOTAN_PUBLIC_API(2,0) XMSS_PrivateKey final : public virtual XMSS_PublicKe
       size_t size() const override
          {
          return XMSS_PublicKey::size() +
-                sizeof(uint64_t) +
+                sizeof(uint32_t) +
                 2 * XMSS_PublicKey::m_xmss_params.element_size();
          }
